+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 - Draw: Adds a new attribute semicolonCSVDelimiter to decide the semicolon is used as the delimiter for exported CSV file. The default value is true.
 - layerTree: Adds a new parameter in portalconfig.tree to decide whether the scale is displayed in the layer tooltip in addition to the name of the corresponding layer
+- print: possibility to edit the output file name in print mode 'plotservice'.
 
 ### Changed